+++ cvat/apps/engine/static/engine/js/logger.js
@@ -363,6 +363,8 @@ var Logger = {
         debugInfo: 24,
         // dumped as "Fit image". There are no additional required fields.
         fitImage: 25,
+        // dumped as "Rotate image". There are no additional required fields.
+        rotateImage: 26,
     },
 
     /**
@@ -526,6 +528,7 @@ var Logger = {
         case this.EventType.changeFrame: return 'Change frame';
         case this.EventType.debugInfo: return 'Debug info';
         case this.EventType.fitImage: return 'Fit image';
+        case this.EventType.rotateImage: return 'Rotate image';
         default: return 'Unknown';
         }
     },